Created new testmap for conveyors and a test class (not any tests yet)
