+++ assets/testmaps/conveyor_tile_test_map.tmx
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<map version="1.2" tiledversion="1.2.3" orientation="orthogonal" renderorder="right-down" width="10" height="10" tilewidth="300" tileheight="300" infinite="0" nextlayerid="2" nextobjectid="1">
+ <tileset firstgid="1" source="../tiles_tileset.tsx"/>
+ <layer id="1" name="Tile Layer 1" width="10" height="10">
+  <data encoding="csv">
+5,5,5,5,5,5,5,5,5,5,
+5,5,5,5,5,5,5,5,5,5,
+5,5,5,5,5,5,5,5,5,5,
+35,5,5,5,5,19,5,5,5,5,
+49,5,5,5,5,13,5,5,5,5,
+5,5,5,5,5,5,5,5,5,5,
+5,5,5,5,5,21,5,21,5,5,
+5,5,5,5,5,21,13,5,5,5,
+50,51,5,5,5,13,22,22,5,5,
+49,52,5,5,5,13,14,14,5,5
+</data>
+ </layer>
+</map>
